### 2019-08-23 #1679 @whedon
[PRE REVIEW]: Python KM: Kramers--Moyal coefficients for stochastic processes
**Submitting author:** @LRydin (<a href="http://orcid.org/0000-0001-5513-0580">Leonardo Rydin Gorjão</a>)
**Repository:** <a href="https://github.com/LRydin/KramersMoyal" target ="_blank">https://github.com/LRydin/KramersMoyal</a>
**Version:** v0.31
