### 2019-08-23 #1679 @kyleniemeyer
@arfon not sure what's going on with the paper here—it builds fine for me locally. Could the presence of `paper.tex` and `paper.pdf` be causing an issue?
