### 2019-08-23 #1679 @kyleniemeyer
@xuanxu I'll give it a shot, but this is a math-heavy submission, and Viviane is out until October.
