### 2019-08-23 #1679 @arfon
> @arfon not sure what's going on with the paper here—it builds fine for me locally. Could the presence of paper.tex and paper.pdf be causing an issue?

Yeah, @LRydin - could you please remove the `paper.tex` file from your repository? It's confusing @whedon.
