### 2019-08-23 #1679 @LRydin
Dear All,

@arfon certainly, I'll remove the .tex. Do understand that the paper.pdf is
generated from that TeX, since I am a novice with Markdown.
