### 2019-08-23 b749f00
At request, .tex is remove. Current paper was generated from this tex
