+++ paper/paper.tex
@@ -1,181 +1,0 @@
-\documentclass[showpacs,showkeys,10pt,onecolumn,superscriptaddress,notitlepage]{revtex4-1}
-
-\usepackage[ansinew]{inputenc}
-\usepackage[T1]{fontenc}
-
-\usepackage{float}
-
-\usepackage{amsmath,amsthm,amssymb,amscd}
-\usepackage{bbm}
-\usepackage{mathtools}
-\usepackage{hyperref}
-
-\bibliographystyle{plain}
-
-\newcommand{\Epi}{\affiliation{Department of Epileptology, University of Bonn, Venusberg Campus 1, 53127~Bonn, Germany}}
-\newcommand{\HISKP}{\affiliation{Helmholtz Institute for Radiation and Nuclear Physics, University of Bonn, Nussallee~14--16, 53115~Bonn, Germany}}
-\newcommand{\THP}{\affiliation{Institute for Theoretical Physics, University of Cologne, 50937 K\"oln, Germany}}
-\newcommand{\FZJ}{\affiliation{Forschungszentrum J\"ulich, Institute for Energy and Climate Research - Systems Analysis and Technology Evaluation (IEK-STE), 52428 J\"ulich, Germany}}
-\newcommand{\Bethe}{\affiliation{Physikalisches Institut and Bethe Center for Theoretical Physics, Universit\"at Bonn, Nussallee 12, 53115 Bonn, Germany}}
-
-
-\frenchspacing
-
-\begin{abstract}
-    This python library allows the numerical calculation of Kramers--Moyal coefficients for stochastic data of any dimension and for any desired order.
-    The library relies on a kernel-based estimation to obtain accurate results for data with a low number of data points, employing a convolution schema to circumvent the traditional drawbacks of binning data, which is especially helpful for larger datasets.
-    The flexibility of the code allows the calculation of the Kramers--Moyal coefficients for data of any dimension, up to any order, without additional parameters, while including a handful of kernels for the estimation, up to user's choice.
-\end{abstract}
-
-\begin{document}
-
-\title{Python KM: Kramers--Moyal coefficients for stochastic processes}
-
-
-\author{Leonardo~Rydin~Gorj\~ao}
-\email[Electronic mail: ]{l.rydin.gorjao@fz-juelich.de}
-\Epi \HISKP \THP \FZJ
-
-\author{Francisco Meirinhos}
-\email[Electronic mail: ]{meirinhos@physik.uni-bonn.de}
-\Bethe
-
-\keywords{Kramers--Moyal coefficents, N-dimensional stochastic processes, Markovian processes, Master equation, Fokker--Planck equation, Ornstein--Uhlenbeck process, Python}
-
-\maketitle
-
-\textbf{Library available at}:  \texttt{\href{https://github.com/LRydin/KramersMoyal}{github.com/LRydin/KramersMoyal}}
-
-\section{Summary}
-
-A general problem for evaluating Markovian stochastic processes is the retrieval of the moments or the Kramers--Moyal coefficients $\mathcal{M}$ from data or time-series.
-The Kramers--Moyal coefficients are derived from an Taylor expansion of the master equation that describes the probability evolution of a Markovian stochastic process.
-
-Given a set of stochastic data, ergodic or quasi-stationary, the extensive literature of stochastic processes awards a set of measures, such as the Kramers--Moyal coefficients or its moments, which link stochastic processes to a probabilistic description of the process or of the family of processes \cite{Risken}.
-Most commonly known is the Fokker--Planck equation or truncated forward Kolmogorov equation, partial differential equations, obtained from the Taylor expansion of the master equation.
-
-
-
[…]
