### 2019-08-23 #1679 @arfon
@LRydin - this pull request should fix your paper: https://github.com/LRydin/KramersMoyal/pull/3 . I'm not sure if this author has an ORDIC but the placeholder one 0000-0000-0000-0000 is failing validation.

The Pandoc rendering engine goes from Markdown (`paper.md`) straight through to pdf. Having a file called `paper.tex` confuses our robot Whedon when compiling the paper :-)
