### 2019-08-23 #1679 @kthyng
Sure, I can be editor.
