### 2019-08-23 #1679 @kthyng
@LRydin Can you look at the list of reviewers (https://bit.ly/joss-reviewers) and select out at least 4 that look like good fits for your submission? Then I will ask them, in search of 2. People not from the list are fine, too.
