### 2019-08-24 #1679 @LRydin
@kthyng <https://github.com/kthyng> Hey Kristen Thyng, I have found four
possible names:

pdebuyl
shibabrat
dawbarton
JoshKarpel

there is a possible other reviewer that has worked in the field and now is
a Professor at the Metropolitan University of Oslo: Pedro Lind,
Pedro.Lind@oslomet.no

Best,
Leonardo

On Fri, 23 Aug 2019 at 22:51, Kristen Thyng <notifications@github.com>
wrote:

> @LRydin <https://github.com/LRydin> Can you look at the list of reviewers
> (https://bit.ly/joss-reviewers) and select out at least 4 that look like
> good fits for your submission? Then I will ask them, in search of 2. People
> not from the list are fine, too.
>
