### 2019-08-24 6f23db1
Added missing powers and eps details to kmc. It seems eps is not used?
