### 2019-08-27 #1679 @dawbarton
Yes, I should be able to do that. I've strayed into this sort of area before, my knowledge is a bit rusty but should be fine.
