### 2019-08-27 #1679 @LRydin
Dear Kristen Thyng,

I am but a PhD student, I don't think I possess at the time the required
knowledge to review papers, and the underlying python coding here, i.e.,
not the fundamental idea, but the improvement of the code, was performed by
Francisco Meirinhos (the second author). I feel he has the skills to review
computationally oriented papers. Maybe he would be a better bet?

I am thankful for the offer, but I don't believe I would have the required
knowledge for such.

Best,
Leonardo Rydin

On Tue, 27 Aug 2019 at 19:49, Kristen Thyng <notifications@github.com>
wrote:

> @LRydin <https://github.com/LRydin> By the way several physics-based
> submissions have just come in, perhaps you would like to be added to the
> list of reviewers for JOSS as more come in? You can sign up easily
> <https://joss.theoj.org/reviewer-signup.html>.
>
