### 2019-08-30 #1693 @whedon
[REVIEW]: Python KM: Kramers--Moyal coefficients for stochastic processes
**Submitting author:** @LRydin (<a href="http://orcid.org/0000-0001-5513-0580">Leonardo Rydin Gorjão</a>)
**Repository:** <a href="https://github.com/LRydin/KramersMoyal" target ="_blank">https://github.com/LRydin/KramersMoyal</a>
**Version:** v0.4
**Reviewer:** @dawbarton , @Shibabrat
**Archive:** <a href="https://doi.org/10.21105/joss.01693" target="_blank">10.21105/joss.01693</a>
