### 2019-08-30 91c6176
Corrected error in binning_test and added bicount_test
