+++ kramersmoyal/kernels.py
@@ -57,6 +57,7 @@ def uniform(x, dims):
     normalization = volume_unit_ball(dims)
     return kernel / normalization
 
+
 _kernels = {epanechnikov, gaussian, uniform}
 
 
