### 2019-09-11 LRydin/KramersMoyal#4 @dawbarton
Errors in the readme [joss.01693]
I'm just going through your submission to JOSS and following through your readme I can't reproduce your results exactly.

1. The code errors since `kernel.epanechnikov` is an unknown name  (it should be `kernels.epanechnikov`).
2. The code errors saying that the first power should be zero.

Fixing those two mistakes I then get the results that I plot using

```
plot(edges[0], kmc[1])  # the first coefficient
```

There is some distortion at the extremes (I presume to be expected due to low numbers of samples?) and it shows the linear trend demonstrated in the readme figures but the values are a factor of 1000 different.

A couple of minor points that don't affect the calculation but are more about style:
1. `y = np.zeros([time.size])` could simply be `y = np.zeros(time.size)`, and
2. `size=[time.size,1]` could simply be `size=time.size`.
