### 2019-09-11 LRydin/KramersMoyal#4 @dawbarton
Also, why is `bins` a numpy array rather than a simple integer? And the same for `powers` - is there a reason it is an array of arrays of integers rather than a simple array of integers?
