### 2019-09-11 LRydin/KramersMoyal#4 @dawbarton
Also, according to the JOSS checklist, you should state the dependencies of the package in your documentation (in this case, your readme since you have no separate documentation). I note that you mention it in the paper but it isn't immediately obvious and JOSS ask for it in the documentation rather than the paper.

In addition to the points already mentioned, I currently don't feel I can tick off

1. Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
     * There is no API documentation as far as I can see. I note there are docstrings but a bit more in the readme would be better. Also when the input is an array, a bit more information about what is expected would be helpful (for example, see my question about `powers` above).
2. Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
    * I didn't spot any of these.
