### 2019-09-11 LRydin/KramersMoyal#5 @dawbarton
Comments on the paper [joss.01693]
Overall the paper reads well but there are a number of typos to be fixed - see the attached PDF. (There is a little to much enthusiasm for commas though.)

[10.21105.joss.01693-annotated.pdf](https://github.com/LRydin/KramersMoyal/files/3601741/10.21105.joss.01693-annotated.pdf)
