### 2019-09-11 #1693 @dawbarton
I've made comments for the authors in https://github.com/LRydin/KramersMoyal/issues/4 and https://github.com/LRydin/KramersMoyal/issues/5
