### 2019-09-12 LRydin/KramersMoyal#5 @LRydin
@dawbarton Thank you for the revisions. There are indeed a few mistakes, I will implement your corrections. The references to the figures and equations are indeed missing. I am unfamiliar with Markdown, so I wrote the paper on TeX and likely the \ref command does work on Markdown. I'll upload a corrected version of the paper to the repo. @fmeirinhos
