### 2019-09-16 LRydin/KramersMoyal#4 @LRydin
@dawbarton Sorry for the late reply. I am working with @fmeirinhos to try and reply to your comments.

We have not generated any documentation since I don't know how to do this at this stage. I was initially working on writing down the documentation inside the .py code, and then extended to a full documentation. As I am not an expert in this, it will certainly take a bit of time.

Regarding the community guidelines, this clearly needs to be addressed. I will confer with @fmeirinhos on how to best set this up.
