### 2019-09-26 38b4c84
Fixed kernels typo in README
