### 2019-09-26 ec7ae2e
Now zeroth power of the km algorithm doesn't have to be specified by the user
