+++ kramersmoyal/binning.py
@@ -21,41 +21,41 @@ def bincount2(x, weights, minlength=0):
     return weights * csr
 
 
-def _get_outer_edges(a, range, bw):
-    """
-    Determine the outer bin edges to use, from either the data or the range
-    argument
-    """
-    if range is not None:
-        first_edge, last_edge = range
-        if first_edge > last_edge:
-            raise ValueError(
-                'max must be larger than min in range parameter.')
-        if not (np.isfinite(first_edge) and np.isfinite(last_edge)):
-            raise ValueError(
-                "supplied range of [{}, {}] is not finite".format(first_edge, last_edge))
-    elif a.size == 0:
-        # handle empty arrays. Can't determine range, so use 0-1.
-        first_edge, last_edge = 0, 1
-    else:
-        first_edge, last_edge = a.min() - bw, a.max() + bw
-        if not (np.isfinite(first_edge) and np.isfinite(last_edge)):
-            raise ValueError(
-                "autodetected range of [{}, {}] is not finite".format(first_edge, last_edge))
-
-    # expand empty range to avoid divide by zero
-    if first_edge == last_edge:
-        first_edge = first_edge - 0.5
-        last_edge = last_edge + 0.5
-
-    return first_edge, last_edge
-
-
 _range = range
 
 
 # An alternative to Numpy's histogramdd, supporting a weights matrix
+# Part of the following code is licensed under the BSD-3 License (from Numpy)
 def histogramdd(sample, bins=10, range=None, normed=None, weights=None, density=None, bw=0.0):
+
+    def _get_outer_edges(a, range, bw):
+        """
+        Determine the outer bin edges to use, from either the data or the range
+        argument
+        """
+        if range is not None:
+            first_edge, last_edge = range
+            if first_edge > last_edge:
+                raise ValueError(
+                    'max must be larger than min in range parameter.')
+            if not (np.isfinite(first_edge) and np.isfinite(last_edge)):
+                raise ValueError(
+                    "supplied range of [{}, {}] is not finite".format(first_edge, last_edge))
+        elif a.size == 0:
+            # handle empty arrays. Can't determine range, so use 0-1.
+            first_edge, last_edge = 0, 1
+        else:
+            first_edge, last_edge = a.min() - bw, a.max() + bw
+            if not (np.isfinite(first_edge) and np.isfinite(last_edge)):
+                raise ValueError(
+                    "autodetected range of [{}, {}] is not finite".format(first_edge, last_edge))
+
+        # expand empty range to avoid divide by zero
+        if first_edge == last_edge:
+            first_edge = first_edge - 0.5
+            last_edge = last_edge + 0.5
+
+        return first_edge, last_edge
 
     try:
         # Sample is an ND-array.
