### 2019-09-27 d4f3d13
Added a Contribution section on the README
