### 2019-09-27 0e5eec0
Corrections to README, setup, and ipynb
