+++ README.md
@@ -15,7 +15,10 @@ from kramersmoyal import km, kernels
 ```
 
 # A one-dimensional stochastic process
-## The theory
+
+A Jupyter notebook with this example can be found [here](/examples/kmc.ipynb)
+
+ ## The theory
 Take for example the well documented one-dimension Ornstein–Uhlenbeck process, also known as Va&#353;&#237;&#269;ek process, see [here](https://en.wikipedia.org/wiki/Ornstein%E2%80%93Uhlenbeck_process). This process is governed by two main parameters: the mean-reverting parameter &theta; and the diffusion parameter &sigma;
 
 ![Ornstein–Uhlenbeck process](/other/OU_eq.png)
@@ -53,7 +56,7 @@ for i in range(1,time.size):
 
 From here we have a plain example of an Ornstein–Uhlenbeck process, always drifting back to zero, due to the mean-reverting drift &theta;. The effect of the noise effect can be seen across the whole trajectory.
 
-![Jump-diffusion process](/other/O-U_plot.png)
+<img src="/other/fig1.png" title="Ornstein–Uhlenbeck process" height="200"/>
 
 ## Using Python KM
 Take the timeseries `y` and lets study the Kramers–Moyal coefficients. For this lets look at the drift and diffusion coefficients of the process, i.e., the first and second Kramers–Moyal coefficients, with an `epanechnikov` kernel
@@ -72,15 +75,101 @@ kmc, edges = km(y, kernel=kernels.epanechnikov, bw=bw, bins=bins, powers=powers)
 ```
 
 This results in
-![Jump-diffusion process](/other/O-U_drift_diffusion.png)
+
+<img src="/other/fig2.png" title="Drift and diffusion terms" height="200"/>
 
 
-# A bivariate jump-diffusion process
-A bivariate jump-diffusion process, a two-dimensional process with higher-order statistical moments, comprises interaction between both the noise terms from the Wiener process ![Wiener process](/other/W.png) and the Poissonian jumps terms ![Poissonian jump process](/other/J.png) across dimensions.
+# A 2-dimensional diffusion process
 
-![Jump-diffusion process](/other/JD_process.png)
+A Jupyter notebook with this example can be found [here](/examples/kmc.ipynb)
 
-An integration scheme as mentioned above allows one to numerically simulate the process.
+A 2-dimensional diffusion process is an process that comprises two ![Wiener process](/other/inline_W.png) and allows for a mixing of these noise terms across its two dimensions.
+
+## Theory
+A two-dimensionan diffusion process takes the form
+
+<img src="/other/2D-diffusion.png" alt="2D-diffusion" title="A 2-dimensional diffusion process" height="60" />
+
+where we will select a set of more complicated parameters obeying
+
+<img src="/other/parameters_2D-diffusion.png" alt="2D-diffusion" title="A 2-dimensional diffusion process" height="70" />
+
+with ![](/other/inline_parameters_2D-diffusion.png).
+
+## Choice of parameters
+As an example, lets take the following set of parameters for the drift vector and diffusion matrix
+
+```python
+# Define the drift vector N
+N = np.array([2.0, 1.0])
+
+# Define the diffusion matrix g
+g = np.array([[0.5, 0.0],[0.0, 0.5]])
+
+# The total integration time
+t_start = 0
+t_final = 2000
+
+# The desired timestep of integration
+delta_t = 0.001
+
+# time array of the process
+time = np.linspace(0, t_final, t_final * int(1 / delta_t))
+```
+## Integrating a 2-dimensional process
[…]
