### 2019-09-27 b8cfa43
Corrections to README
