+++ kramersmoyal/kmc.py
@@ -7,7 +7,8 @@ from .kernels import silvermans_rule, epanechnikov, _kernels
 
 
 def km(timeseries: np.ndarray, bins: np.ndarray, powers: np.ndarray,
-        kernel=None, bw=None, tol=1e-10, conv_method='auto'):
+        kernel: callable=None, bw: float=None, tol: float=1e-10,
+        conv_method: str='auto'):
     """
     Estimates Kramers-Moyal coefficients from a timeseries using a kernel
     estimator method.
@@ -15,19 +16,35 @@ def km(timeseries: np.ndarray, bins: np.ndarray, powers: np.ndarray,
     Parameters
     ----------
     timeseries: np.ndarray
-        The D-dimensional timeseries (N, D)
+        The D-dimensional timeseries (N, D). The timeseries of length N and
+        dimensions D.
 
     bins: np.ndarray
-        The number of bins for each dimension
+        The number of bins for each dimension. This is the underlying space for
+        the Kramers-Moyal coefficients. In 1-dimension a choice as
+            bins = np.array([6000])
+        is recommended. In 2-dimensions
+            bins = np.array([300,300])
+        is recommended
 
     powers: np.ndarray
-        Powers for the operation of calculating the Kramers--Moyal coefficients
+        Powers for the operation of calculating the Kramers-Moyal coefficients,
+        which need to match dimensions of the timeseries. In 1-dimension the
+        first four Kramers-Moyal coefficients can be found via
+            powers = np.array([0],[1],[2],[3],[4])
+        In 2 dimensions take into account each dimension, as
+            powers = np.array([0,0],[0,1],[1,0],[1,1],[0,2],[2,0],[2,2],
+                              [0,3],[3,0],[3,3],[0,4],[4,0],[4,4])
 
     kernel: callable
-        Kernel used to convolute with the Kramers--Moyal coefficients
+        Kernel used to convolute with the Kramers-Moyal coefficients. To select
+        for example a Gaussian kernel use
+            kernel = kernels.epanechnikov
+        If None the Epanechnikov kernel will be used
 
     bw: float
-        Desired bandwidth of the kernel
+        Desired bandwidth of the kernel. A value of 1 occupies the full space of
+        the bin space. Recommended are values 0.005 < bw < 0.2
 
     tol: float
         Round to zero absolute values smaller than `tol`, after the convolutions
@@ -39,10 +56,13 @@ def km(timeseries: np.ndarray, bins: np.ndarray, powers: np.ndarray,
     Returns
     -------
     kmc: np.ndarray
-        The calculated Kramers-Moyal coefficients
+        The calculated Kramers-Moyal coefficients in accordance to the
+        timeseries dimensions in (D,bins.shape) shape. To extract the selected
+        orders of the kmc, use kmc[i,...], with i the order according to powers
 
     edges: np.ndarray
-        The bin edges of the calculated Kramers-Moyal coefficients
+        The bin edges with shape (D,bins.shape) of the calculated Kramers-Moyal
+        coefficients
     """
     timeseries = np.asarray_chkfinite(timeseries, dtype=float)
     if len(timeseries.shape) == 1:
@@ -82,6 +102,10 @@ def km(timeseries: np.ndarray, bins: np.ndarray, powers: np.ndarray,
 
 def _km(timeseries: np.ndarray, bins: np.ndarray, powers: np.ndarray,
         kernel: callable, bw: float, tol: float, conv_method: str):
+    """
+    Helper function for km that does the heavy lifting and actually estimates
+    the Kramers-Moyal coefficients from the timeseries.
+    """
[…]
