+++ README.md
@@ -21,7 +21,7 @@ A Jupyter notebook with this example can be found [here](/examples/kmc.ipynb)
  ## The theory
 Take for example the well documented one-dimension Ornstein–Uhlenbeck process, also known as Va&#353;&#237;&#269;ek process, see [here](https://en.wikipedia.org/wiki/Ornstein%E2%80%93Uhlenbeck_process). This process is governed by two main parameters: the mean-reverting parameter &theta; and the diffusion parameter &sigma;
 
-![Ornstein–Uhlenbeck process](/other/OU_eq.png)
+<img src="/other/OU_eq.png" title="Ornstein–Uhlenbeck process" height="25"/>
 
 which can be solved in various ways. For our purposes, recall that the drift coefficient, i.e., the first-order Kramers–Moyal coefficient, is given by ![first-order Kramers–Moyal coefficient of an Ornstein–Uhlenbeck process](/other/KM_1.png) and the second-order Kramers–Moyal coefficient is ![second-order Kramers–Moyal coefficient of an Ornstein–Uhlenbeck process](/other/KM_2.png), i.e., the diffusion.
 
@@ -37,8 +37,8 @@ t_final = 500
 delta_t = 0.001
 
 # The parameters theta and sigma
-theta = .3
-sigma = .1
+theta = 0.3
+sigma = 0.1
 
 # The time array of the trajectory
 time = np.arange(0,t_final,delta_t)
@@ -83,39 +83,37 @@ This results in
 
 A Jupyter notebook with this example can be found [here](/examples/kmc.ipynb)
 
-A 2-dimensional diffusion process is an process that comprises two ![Wiener process](/other/inline_W.png) and allows for a mixing of these noise terms across its two dimensions.
-
 ## Theory
+
+A 2-dimensional diffusion process is an process that comprises two ![Wiener process](/other/inline_W.png) and allows for a mixing of these noise terms across its two dimensions.
 A two-dimensionan diffusion process takes the form
 
 <img src="/other/2D-diffusion.png" alt="2D-diffusion" title="A 2-dimensional diffusion process" height="60" />
 
 where we will select a set of more complicated parameters obeying
 
-<img src="/other/parameters_2D-diffusion.png" alt="2D-diffusion" title="A 2-dimensional diffusion process" height="70" />
+<img src="/other/parameters_2D-diffusion.png" alt="2D-diffusion" title="Specific parameters for the diffusion process" height="70" />
 
-with ![](/other/inline_parameters_2D-diffusion.png).
+with ![](/other/inline_parameters_2D-diffusion_1.png) and ![](/other/inline_parameters_2D-diffusion_2.png).
 
 ## Choice of parameters
 As an example, lets take the following set of parameters for the drift vector and diffusion matrix
 
 ```python
+# integration time and time sampling
+t_final = 2000
+delta_t = 0.001
+
 # Define the drift vector N
 N = np.array([2.0, 1.0])
 
 # Define the diffusion matrix g
 g = np.array([[0.5, 0.0],[0.0, 0.5]])
 
-# The total integration time
-t_start = 0
-t_final = 2000
-
-# The desired timestep of integration
-delta_t = 0.001
-
-# time array of the process
-time = np.linspace(0, t_final, t_final * int(1 / delta_t))
+# The time array of the trajectory
+time = np.arange(0,t_final,delta_t)
 ```
+
 ## Integrating a 2-dimensional process
[…]
