### 2019-09-29 4a87738
Added literature, change history of project, corrections of typos
