+++ README.md
@@ -2,6 +2,7 @@
 Python KM is a  python package designed to obtain the Kramers–Moyal coefficients, or conditional moments, from stochastic data of any dimension. It employs kernel density estimations, instead of a histogram approach, to ensure better results for low number of points as well as allowing better fitting of the results
 
 # Changelog
+- Version 0.32 - Adding 2 kernels: Triagular and Quartic and extentind the documentation and examples.
 - Version 0.31 - Corrections to the fft triming after convolution
 - Version 0.3 - The major breakthrough: Calculates the Kramers–Moyal coefficients for data of any dimension
 - Version 0.2 - Introducing convolutions and `gaussian` and `uniform` kernels. Major speed up in the calculations.
@@ -25,8 +26,8 @@ Take for example the well documented one-dimension Ornstein–Uhlenbeck process,
 
 which can be solved in various ways. For our purposes, recall that the drift coefficient, i.e., the first-order Kramers–Moyal coefficient, is given by ![first-order Kramers–Moyal coefficient of an Ornstein–Uhlenbeck process](/other/KM_1.png) and the second-order Kramers–Moyal coefficient is ![second-order Kramers–Moyal coefficient of an Ornstein–Uhlenbeck process](/other/KM_2.png), i.e., the diffusion.
 
-Generate an exemplary Ornstein–Uhlenbeck process with your favorite integrator, e.g., the [Euler–Maruyama](https://en.wikipedia.org/wiki/Euler%E2%80%93Maruyama_method) or with a more powerful tool from [`JiTCSDE`](https://github.com/neurophysik/jitcsde) found on GitHub.
-For this example lets take &theta;=.3 and &sigma;=.1, over a total time of 500 units, with a sampling of 1000 Hertz, and from the generated data series retrieve the two parameters, the drift &theta; and diffusion &sigma;.
+Generate an exemplary Ornstein–Uhlenbeck process with your favourite integrator, e.g., the [Euler–Maruyama](https://en.wikipedia.org/wiki/Euler%E2%80%93Maruyama_method) or with a more powerful tool from [`JiTCSDE`](https://github.com/neurophysik/jitcsde) found on GitHub.
+For this example let's take &theta;=.3 and &sigma;=.1, over a total time of 500 units, with a sampling of 1000 Hertz, and from the generated data series retrieve the two parameters, the drift &theta; and diffusion &sigma;.
 
 ## Integrating an Ornstein–Uhlenbeck process
 Here is a short code on generating a Ornstein–Uhlenbeck stochastic trajectory with a simple Euler–Maruyama integration method
@@ -41,13 +42,13 @@ theta = 0.3
 sigma = 0.1
 
 # The time array of the trajectory
-time = np.arange(0,t_final,delta_t)
+time = np.arange(0, t_final, delta_t)
 
 # Initialise the array y
 y = np.zeros(time.size)
 
-#Generate a Wiener process
-dw = np.random.normal(loc=0, scale=np.sqrt(delta_t),size=time.size)
+# Generate a Wiener process
+dw = np.random.normal(loc=0, scale=np.sqrt(delta_t), size=time.size)
 
 # Integrate the process
 for i in range(1,time.size):
@@ -59,7 +60,7 @@ From here we have a plain example of an Ornstein–Uhlenbeck process, always dri
 <img src="/other/fig1.png" title="Ornstein–Uhlenbeck process" height="200"/>
 
 ## Using Python KM
-Take the timeseries `y` and lets study the Kramers–Moyal coefficients. For this lets look at the drift and diffusion coefficients of the process, i.e., the first and second Kramers–Moyal coefficients, with an `epanechnikov` kernel
[…]
