### 2019-09-29 08b144e
Adding setup.py for installation (ongoing)
