### 2019-09-29 5b03d18
Adding more documentation and 2 new kernels
