+++ kramersmoyal/__init__.py
@@ -1,3 +1,5 @@
 from .kernels import epanechnikov, gaussian, uniform
 from .kernels import silvermans_rule
 from .kmc import km
+
+name = "kramersmoyal"
