+++ kramersmoyal/kernels.py
@@ -8,6 +8,12 @@ def kernel(kernel_func):
     """
     Transforms a kernel function into a scaled kernel function
     (for a certain bandwidth `bw`)
+
+    Currently implemented kernels are:
+        Epanechnikov, Gaussian, Uniform, Triangular, Quartic
+
+    For a good overview of various kernels see
+    https://en.wikipedia.org/wiki/Kernel_(statistics)
     """
     @wraps(kernel_func)  # just for naming
     def decorated(x, bw):
@@ -23,47 +29,83 @@ def kernel(kernel_func):
     return decorated
 
 
-def volume_unit_ball(dims):
+def volume_unit_ball(dims: int) -> float:
+    """
+    Returns the volume of a unit ball in dimensions dims
+    """
     return np.pi ** (dims / 2.0) / gamma(dims / 2.0 + 1.0)
 
-
 @kernel
-def epanechnikov(x, dims):
+def epanechnikov(x: np.ndarray, dims: int) -> np.ndarray:
+    """
+    The Epanechnikov kernel in dimensions dims
+    """
     x2 = (x ** 2)
     mask = x2 < 1.0
     kernel = np.zeros_like(x)
     kernel[mask] = (1.0 - x2[mask])
-    normalization = 2.0 / (dims + 2.0) * volume_unit_ball(dims)
-    return kernel / normalization
-
+    normalisation = 2.0 / (dims + 2.0) * volume_unit_ball(dims)
+    return kernel / normalisation
 
 @kernel
-def gaussian(x, dims):
+def gaussian(x: np.ndarray, dims: int) -> np.ndarray:
+    """
+    Gaussian kernel in dimensions dims
+    """
     def gaussian_integral(n):
         if n % 2 == 0:
             return np.sqrt(np.pi * 2) * factorial2(n - 1) / 2
         elif n % 2 == 1:
             return np.sqrt(np.pi * 2) * factorial2(n - 1) * norm.pdf(0)
     kernel = np.exp(-x ** 2 / 2.0)
-    normalization = dims * gaussian_integral(dims - 1) * volume_unit_ball(dims)
-    return kernel / normalization
-
+    normalisation = dims * gaussian_integral(dims - 1) * volume_unit_ball(dims)
+    return kernel / normalisation
 
 @kernel
-def uniform(x, dims):
+def uniform(x: np.ndarray, dims: int) -> np.ndarray:
+    """
+    Uniform, or rectangular kernel in dimensions dims
+    """
     mask = x < 1.0
     kernel = np.zeros_like(x)
     kernel[mask] = 1.0
-    normalization = volume_unit_ball(dims)
-    return kernel / normalization
+    normalisation = volume_unit_ball(dims)
+    return kernel / normalisation
+
+@kernel
+def triagular(x: np.ndarray, dims: int) -> np.ndarray:
+    """
+    Triagular kernel in dimensions dims
+    """
+    mask = x < 1.0
+    kernel = np.zeros_like(x)
+    kernel[mask] = 1.0 - np.abs(x[mask])
+    normalisation = volume_unit_ball(dims) / 2.0
+    return kernel / normalisation
 
 
-_kernels = {epanechnikov, gaussian, uniform}
+@kernel
+def quartic(x: np.ndarray, dims: int) -> np.ndarray:
+    """
+    Quartic, or biweight kernel in dimensions dims
+    """
+    x2 = (x ** 2)
+    mask = x2 < 1.0
+    kernel = np.zeros_like(x)
+    kernel[mask] = ((1.0 - x2[mask]) ** 2)
+    normalisation = 2.0 / (dims + 2.0) * volume_unit_ball(dims)
+    return kernel / normalisation
 
 
[…]
