+++ setup.py
@@ -4,13 +4,11 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="km",
+    name="kramersmoyal",
     version="0.3.2",
     author="Leonardo Rydin Gorjao and Francisco Meirinhos",
     author_email="leonardo.rydin@gmail.com",
-    description="Calculate Kramers-Moyal coefficients\
-                 for stochastic process of any dimension,\
-                 up to any order.",
+    description="Calculate Kramers-Moyal coefficients for stochastic process of any dimension, up to any order.",
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/LRydin/KramersMoyal",
