### 2019-09-29 66815c3
Updates to Paper in reply to issue: Comments on the paper [joss.01693] #5
