+++ paper/paper.md
@@ -37,51 +37,51 @@ The Kramers--Moyal coefficients are derived from an Taylor expansion of the mast
 Given a set of stochastic data, ergodic or quasi-stationary, the extensive literature of stochastic processes awards a set of measures, such as the Kramers--Moyal coefficients or its moments, which link stochastic processes to a probabilistic description of the process or of the family of processes [@Risken].
 Most commonly known is the Fokker--Planck equation or truncated forward Kolmogorov equation, partial differential equations, obtained from the Taylor expansion of the master equation.
 
-
-
 Of particular relevance is the growing evidence that real-world data displays higher-order ($n>2$) Kramers--Moyal coefficients, which has a two-fold consequence: The common truncation at third order of the forward Kolmogorov equation, giving rise to the Fokker--Planck equation, is no longer valid.
 The evidence of higher-order ($n>2$) Kramers--Moyal coefficients in recorded data thus invalidates the aforementioned common argument for truncation, based on Pawula's theorem and thus rendering the Fokker--Planck description unjustified [@Tabar].
-A clear and common example is the presence of discontinuous jumps in data[@Anvari,Sahalia], which can give rise to higher-order Kramers--Moyal coefficients, and are evidenced in [@Rydin] and references within.
+A clear and common example is the presence of discontinuous jumps in data[@Anvari] [@Sahalia], which can give rise to higher-order Kramers--Moyal coefficients, and are evidenced in [@Rydin] and references within.
 
 
-Calculating the moments or Kramers--Moyal coefficients strickly from data can be computationally heavy for long data series and is prone to be innacurate especially where the density of data points is scarce, e.g. usually at the boundaries on the domain of the process.
+Calculating the moments or Kramers--Moyal coefficients stricly from data can be computationally heavy for long data series and is prone to be innacurate especially where the density of data points is scarce, e.g. usually at the boundaries on the domain of the process.
 The most straightforward approach is to perform a histogram-based estimation to evaluate the moments of the system at hand.
 This has two main drawbacks: it requires a discrete space of examination of the process, and is shown to be less accurate than using kernel-based estimators [@Lamouroux].
 
 This library is based on a kernel-based estimation, which allows for more robust results given both a wider range of possible kernel shapes to perform the calculation, as well as retrieving the results in a non-binned coordinate space, unlike histogram regressions [@Silverman].
-It further employs a convolution of the series in studied with the selected kernel, circumventing the computational issue of sequential array summation, the most common bottleneck in integration time and computer memory.
+It further employs a convolution of the time series with the selected kernel, circumventing the computational issue of sequential array summation, the most common bottleneck in integration time and computer memory.
 
-The package presented here comprises a manifold of options: A general open-source toolbox for the calculation of Kramers--Moyal coefficients for any given data series of any dimension and to any order, with a selection of commonly used kernel estimators.
[…]
