+++ paper/bib.bib
@@ -30,7 +30,7 @@ URL = {https://doi.org/10.1137/1114019},
 }
 
 @book{Risken,
-  title={The Fokker--{P}lanck {E}quation},
+  title={The {F}okker--{P}lanck {E}quation},
   author={{Risken, H.}},
   booktitle={The {F}okker--{P}lanck {E}quation},
   year={1996},
@@ -111,9 +111,19 @@ URL = {https://doi.org/10.1063/1.5019320}
 }
 
 @article{Rydin,
-  title={Analysis of Bivariate Jump-Diffusion Processes},
-  author={{Rydin Gorj{\~a}o, L., Heysel, J., Lehnertz, K. and Tabar, M. R. R.}},
+  title={Analysis and data-driven reconstruction of bivariate jump-diffusion processes},
+  author={{Gorj{\~a}o, L. R., Heysel, J., Lehnertz, K. and Tabar, M. R. R.}},
   journal={arXiv preprint arXiv:1907.05371},
   year={2019},
-  doi={https://arxiv.org/abs/1907.05371}
+  URL={https://arxiv.org/abs/1907.05371}
+}
+
+@ARTICLE{rinn2016,
+  author = {{Rinn, P., Lind, P., W\"achter, M., and Peinke, J.}},
+  title = {The {L}angevin Approach: An {R} Package for Modeling {M}arkov Processes},
+  journal = {Journal of Open Research Software},
+  year = {2016},
+  volume = {4},
+  pages = {e34},
+  doi = {10.5334/jors.123}
 }
