### 2019-09-29 LRydin/KramersMoyal#5 @LRydin
Dear @dawbarton,

I have implemented you corrections (and asked whedon to generate a new paper.pdf).
https://github.com/openjournals/joss-papers/blob/joss.01693/joss.01693/10.21105.joss.01693.pdf

The indicated annotations you had where inserted.

There are two points I noticed and corrected simultaneously:
- The figure captions I had on my original TeX were never printed by whedon initially (given my lack of knowledge of Markdown). I have somehow found a way to introduce them in a table format. Would you be so kind to glance over the paper once more? (I'm not sure captions for figures are a standard at JOSS)
- I have come across a very relevant package for `R` that tackles similar issues of stochastic processes, that I feel should be cited in our paper, given the relevance of its content. It can calculate Kramers-Moyal coefficients in 1 and 2 dimensions (as far as I understood). It is referred to as [@Rinn2016] and the following was added: "There exists a library to retrieve Kramers--Moyal coefficents in `R` [@rinn2016]."
(https://openresearchsoftware.metajnl.com/article/10.5334/jors.123/)
[the bib.file is therefore also updated]

@kthyng I believe as an editor this information is relevant for you as well.

Best regards,
Leonardo Rydin
@fmeirinhos
