### 2019-09-30 80b35d4
Added a relevant remark on normalisation and another reference
