+++ README.md
@@ -2,7 +2,7 @@
 Python KM is a  python package designed to obtain the Kramers–Moyal coefficients, or conditional moments, from stochastic data of any dimension. It employs kernel density estimations, instead of a histogram approach, to ensure better results for low number of points as well as allowing better fitting of the results
 
 # Changelog
-- Version 0.32 - Adding 2 kernels: Triagular and Quartic and extentind the documentation and examples.
+- Version 0.32 - Adding 2 kernels: `triagular` and `quartic` and extenting the documentation and examples.
 - Version 0.31 - Corrections to the fft triming after convolution
 - Version 0.3 - The major breakthrough: Calculates the Kramers–Moyal coefficients for data of any dimension
 - Version 0.2 - Introducing convolutions and `gaussian` and `uniform` kernels. Major speed up in the calculations.
@@ -24,7 +24,7 @@ Take for example the well documented one-dimension Ornstein–Uhlenbeck process,
 
 <img src="/other/OU_eq.png" title="Ornstein–Uhlenbeck process" height="25"/>
 
-which can be solved in various ways. For our purposes, recall that the drift coefficient, i.e., the first-order Kramers–Moyal coefficient, is given by ![first-order Kramers–Moyal coefficient of an Ornstein–Uhlenbeck process](/other/inline_KM_1.png) and the second-order Kramers–Moyal coefficient is ![second-order Kramers–Moyal coefficient of an Ornstein–Uhlenbeck process](/other/inline_KM_2.png), i.e., the diffusion.
+which can be solved in various ways. For our purposes, recall that the drift coefficient, i.e., the first-order Kramers–Moyal coefficient, is given by ![](/other/inline_KM_1.png) and the second-order Kramers–Moyal coefficient is ![](/other/inline_KM_2.png), i.e., the diffusion.
 
 Generate an exemplary Ornstein–Uhlenbeck process with your favourite integrator, e.g., the [Euler–Maruyama](https://en.wikipedia.org/wiki/Euler%E2%80%93Maruyama_method) or with a more powerful tool from [`JiTCSDE`](https://github.com/neurophysik/jitcsde) found on GitHub.
 For this example let's take &theta;=.3 and &sigma;=.1, over a total time of 500 units, with a sampling of 1000 Hertz, and from the generated data series retrieve the two parameters, the drift &theta; and diffusion &sigma;.
@@ -48,7 +48,7 @@ time = np.arange(0, t_final, delta_t)
 y = np.zeros(time.size)
 
 # Generate a Wiener process
-dw = np.random.normal(loc=0, scale=np.sqrt(delta_t), size=time.size)
+dw = np.random.normal(loc = 0, scale = np.sqrt(delta_t), size = time.size)
 
 # Integrate the process
 for i in range(1,time.size):
@@ -66,28 +66,32 @@ Take the timeseries `y` and let's study the Kramers–Moyal coefficients. For th
 bins = np.array([5000])
 
 # Choose powers to calculate
-powers = np.array([[1],[2]])
+powers = np.array([[1], [2]])
 
 # Choose your desired bandwith
-bw = .15
+bw = 0.15
 
 # The kmc holds the results, where edges holds the binning space
-kmc, edges = km(y, kernel=kernels.epanechnikov, bw=bw, bins=bins, powers=powers)
+kmc, edges = km(y, kernel = kernels.epanechnikov, bw = bw, bins = bins, powers = powers)
 ```
 
 This results in
 
-<img src="/other/fig2.png" title="Drift and diffusion terms" height="200"/>
+<img src="/other/fig2.png" title="Drift and diffusion terms of an Ornstein–Uhlenbeck process" height="200"/>
 
[…]
