### 2019-09-30 LRydin/KramersMoyal#4 @fmeirinhos
@dawbarton

Thank you for the revisions.

Here’s the changelog resulting from your remarks:

> The code errors since kernel.epanechnikov is an unknown name (it should be kernels.epanechnikov).

Corrected.

> The code errors saying that the first power should be zero.

The core algorithm needs the first power to be zero, but it now adds it automatically for internal use if the user hasn’t specified it.

> There is some distortion at the extremes (I presume to be expected due to low numbers of samples?) and it shows the linear trend demonstrated in the readme figures but the values are a factor of 1000 different.

This was a mismatch in the README, since the code wasn’t accounting for normalisation over the time steps; something that should be done for a correct physical interpretation, but not related with the algorithm. Corrected.

> A couple of minor points that don't affect the calculation but are more about style:
	• y = np.zeros([time.size]) could simply be y = np.zeros(time.size), and
	• size=[time.size,1] could simply be size=time.size.

Changed.

> Also, why is bins a numpy array rather than a simple integer? And the same for powers - is there a reason it is an array of arrays of integers rather than a simple array of integers?

Regarding bins, this is because the algorithm works for timeseries in n-dimensions, hence requiring an array of bins, each element specifying the binning in each dimension.
The same logic applies to powers. The user needs to specify the exponent in each direction. However, for the 1D case the user can also pass an array of powers, which then internally gets reshaped into a `(-1, 1)` form.
