### 2019-10-01 7d94e2c
Adding dependencies explicitly on the readme and the pip from TestPyPI
