### 2019-10-07 LRydin/KramersMoyal#5 @kthyng
Hi @LRydin! Some fixed for your references:

- [ ] In bibtex, author names should be separated by "and". In particular, your entry for scikitlearn needs "and" between authors instead of commas. Also your entry for Lehnertz and others. You can see examples here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
- [ ] Please also check the spacing around your references in the paper and whether they have parentheses or not. If you refer to them directly inline, they should not have parentheses. To have parentheses in your reference, use [@referencename] and to not have them, use @referencename.
