### 2019-10-07 0b8a714
Update to bib and paper according to comments by @kthyng (Comments on the paper [joss.01693] #5)
