+++ paper/bib.bib
@@ -1,129 +1,126 @@
 @article{Epanechnikov,
-author = {{Epanechnikov, V.}},
-title = {Non-Parametric Estimation of a Multivariate Probability Density},
-journal = {Theory of Probability \& Its Applications},
-volume = {14},
-number = {1},
-pages = {153-158},
-year = {1969},
-doi = {10.1137/1114019},
-URL = {https://doi.org/10.1137/1114019},
-}
-
-@book{Silverman,
-  title={Density estimation for statistics and data analysis},
-  author={{Silverman, B. W.}},
-  year={2018},
-  publisher={Routledge, New York},
-  doi={10.1201/9781315140919},
-  ISBN={9781315140919}
+  title     = {Non-Parametric Estimation of a Multivariate Probability Density},
+  author    = {Epanechnikov, V.},
+  journal   = {Theory of Probability \& Its Applications},
+  volume    = {14},
+  number    = {1},
+  pages     = {153-158},
+  year      = {1969},
+  doi       = {10.1137/1114019},
+  URL       = {https://doi.org/10.1137/1114019},
 }
 
 @article{scikitlearn,
- title={Scikit-learn: Machine Learning in {P}ython},
- author={{Pedregosa, F., Varoquaux, G., Gramfort, A., Michel, V., Thirion, B., Grisel, O., Blondel, M., Prettenhofer, P., Weiss, R., Dubourg, V., Vanderplas, J., Passos, A., Cournapeau, D., Brucher, M., Perrot, M. and Duchesnay, E.}},
- journal={Journal of Machine Learning Research},
- volume={12},
- pages={2825--2830},
- year={2011},
- url={http://www.jmlr.org/papers/v12/pedregosa11a.html}
-}
-
-@book{Risken,
-  title={The {F}okker--{P}lanck {E}quation},
-  author={{Risken, H.}},
-  booktitle={The {F}okker--{P}lanck {E}quation},
-  year={1996},
-  publisher={Springer, Berlin},
-  doi={10.1007/978-3-642-61544-3},
-  isbn={978-3-540-61530-9}
+  title     = {Scikit-learn: Machine Learning in {P}ython},
+  author    = {Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V. and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P. and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+  journal   = {Journal of Machine Learning Research},
+  volume    = {12},
+  pages     = {2825--2830},
+  year      = {2011},
+  url       = {http://www.jmlr.org/papers/v12/pedregosa11a.html}
 }
 
 @article{Lehnertz,
-  title={Characterizing abrupt transitions in stochastic dynamics},
-  author={{Lehnertz, K., Zabawa, L. and Tabar, M. R. R.}},
-  journal={New Journal of Physics},
-  volume={20},
-  number={11},
-  pages={113043},
-  year={2018},
-  doi={10.1088/1367-2630/aaf0d7},
-  url={https://doi.org/10.1088%2F1367-2630%2Faaf0d7}
+  title     = {Characterizing abrupt transitions in stochastic dynamics},
+  author    = {Lehnertz, K. and Zabawa, L. and Tabar, M. R. R.},
+  journal   = {New Journal of Physics},
+  volume    = {20},
+  number    = {11},
+  pages     = {113043},
+  year      = {2018},
+  doi       = {10.1088/1367-2630/aaf0d7},
+  url       = {https://doi.org/10.1088%2F1367-2630%2Faaf0d7}
 }
 
 @article{Anvari,
[…]
