### 2019-10-07 LRydin/KramersMoyal#5 @LRydin
Dear @kthyng,

I have corrected the bib file and respectively the paper according to the two points you just mentioned. Thank you.

Best,
Leonardo Rydin
@fmeirinhos
