### 2019-10-09 #1693 @Shibabrat
@kthyng It's going slowly since I'm also learning the theory behind the package to be sure my feedback is useful but I'll speed-up. Apologies for the delay.
