### 2019-10-17 LRydin/KramersMoyal#5 @kthyng
Hi @LRydin, great! Still have some changes:

* 3rd paragraph, need space before reference after "... jumps in data"
* 2 "and"s in your reference called "rinn2016".

Rest looks ok but I'll read it more carefully later.
