### 2019-10-17 LRydin/KramersMoyal#5 @kthyng
It's helpful to reference the [paper review issue](https://github.com/openjournals/joss-reviews/issues/1693) from here too.
