### 2019-10-17 LRydin/KramersMoyal#4 @kthyng
@dawbarton Have your concerns here been addressed?
