### 2019-10-17 LRydin/KramersMoyal#5 @LRydin
Hey @kthyng,

I will attend to these right away, thank you.

Best,
Leonardo Rydin
@fmeirinhos

P.S.: I'll refer to the paper review issue
