### 2019-10-17 c428133
Minor details on the paper and bib, as mentioned in: Comments on the paper [joss.01693] #5
