+++ paper/bib.bib
@@ -89,7 +89,7 @@
 
 @article{rinn2016,
   title     = {The {L}angevin Approach: An {R} Package for Modeling {M}arkov Processes},
-  author    = {Rinn, P. and Lind, P. and W{\"a}chter, M. and and Peinke, J.},
+  author    = {Rinn, P. and Lind, P. and W{\"a}chter, M. and Peinke, J.},
   journal   = {Journal of Open Research Software},
   year      = {2016},
   volume    = {4},
