### 2019-10-17 LRydin/KramersMoyal#5 @LRydin
As indicated, we corrected the [above](https://github.com/LRydin/KramersMoyal/issues/5#issuecomment-543263780) comments by @kthyng

Best,
Leonardo Rydin
@fmeirinhos
Referring the [review issue](https://github.com/openjournals/joss-reviews/issues/1693)
