### 2019-10-17 LRydin/KramersMoyal#5 @kthyng
looking good!
