### 2019-10-21 LRydin/KramersMoyal#4 @dawbarton
Yes, I'm happy.
