### 2019-11-16 0d16222
Adding .travis.yml for code testing
