+++ kramersmoyal/__init__.py
@@ -1,5 +1,4 @@
-from .kernels import epanechnikov, gaussian, uniform
-from .kernels import silvermans_rule
+from .kernels import epanechnikov, gaussian, uniform, silvermans_rule
 from .kmc import km
 
 name = "kramersmoyal"
