+++ .travis.yml
@@ -1,0 +1,19 @@
+language: python
+python:
+  - "3.6"
+  - "3.7"
+# command to install dependencies
+install:
+  - export PYTHONPATH=$PYTHONPATH:$(pwd)
+  - pip install -r requirements.txt
+  - pip install -e .
+# command to run tests
+script:
+  # - python kramersmoyal/__init__.py
+  - python kramersmoyal/binning.py
+  - python kramersmoyal/kernels.py
+  # - python kramersmoyal/kmc.py
+
+# no emails, I can check it myself, thank you
+notifications:
+  email: false
