### 2019-11-16 9abc316
Adding travis badge
