### 2019-11-22 LRydin/KramersMoyal#7 @Shibabrat
Community guidelines, code of conduct, issue and PR template
One of the JOSS requirements is to have community guidelines and code of conduct for interacting with the project. Authors can link pointers from the README to options under insights about:

Contribute to the software
Seek support for usage (authors have added this in README and can link it to the guidelines)
Report issues or PR (if the authors think users need to report in a specific style)

There are helpful templates created by GitHub to do these in Insights from the bar just under the repository name.

(Part of openjournals/joss-reviews#1693)
