### 2019-11-25 #1693 @Shibabrat
@kthyng I've gone through the recent version of the software and it satisfies all the requirements, as far as I can access. The only remaining item is the community guidelines issue.
