### 2019-11-25 #1693 @LRydin
Dear @kthyng ,

great news :). I will once more go over the paper, bib, etc, and I will revised with my former and present supervisor which funding applies (I believe as is should be correct, but I want to double check).

I still want to add some docs before the publication is out (as in readthedocs), if that is okay with you. If you feel the publication can be made available before, that is of course also possible.

I will attend to the community guidelines and return to you once it is sorted.

Best,
Leonardo Rydin
@fmeirinhos
