### 2019-11-27 7d8d7a5
First steps at adding readthedocs. Good that rst allows for math! (glances at markdown)
