### 2019-11-27 4bc2508
Adding conduct. This merits an update of the version. We are now at 0.4!
