+++ contributions.md
@@ -1,0 +1,25 @@
+# Contributions
+KramersMoyal has lived so far out of collaborations. We welcome reviews and
+ideas from everyone. If you want to share your ideas or report a bug, open an
+[issue](https://github.com/LRydin/KramersMoyal/issues) here on GitHub, or
+contact us directly: [leonardo.rydin"at"gmail.com](mailto:leonardo.rydin@gmail.com)
+
+We are happy to have you join us in furthering this project, helping us improve
+our code and documentation, make it more user-friendly, broadening its
+applicability, or devising other methodologies of implementation.
+
+If you need any help at understanding the code or theory behind it, or you wish
+to implement it in your own project, contact us, we are here to help.
+
+# Conduct of Fairness
+This package is a research-oriented project and abides to a strict conduct of
+fairness.
+We do not discriminate or accept to partake in any discriminatory acts, either
+against gender, gender identity and expression, sexual orientation, disability,
+personal appearance, ethnicity, race, age, or religion.
+We, as the authors of the package, reserve the right to remove any indication
+from third parties or other contributing members that display behaviours
+directed against the aforementioned.
+Posts, comments, issues raised, or harassment made online or elsewhere that does
+not abide to this conduct of fairness will be dealt with swiftly, deleted or
+edited, and reported when deemed threatening to others or ourselves.
