### 2019-11-27 LRydin/KramersMoyal#7 @LRydin
Dear @Shibabrat and @kthyng,

I have added a Conduct of Fairness, plus a few lines on how to contact us. This includes the contributions.md file.
I've made the first steps in adding the ReadTheDocs (I'll add the content from the readme there and more extensive explanations on the binning and bandwidth).

Could you check if this is what is required? We don't have any specific "report issues" requirements, as in, a style or alike. If there are questions, we are happy to try to help and solve whatever is needed.

Refer: [openjournals/joss-reviews#1693](https://github.com/openjournals/joss-reviews/issues/1693)

Best,
Leonardo Rydin
@fmeirinhos
