### 2019-11-27 #1693 @LRydin
Dear @kthyng ,

I've commented on the open issue [#7](https://github.com/LRydin/KramersMoyal/issues/7), added a Conduct of Fairness, and address a bit more the Contributions. I don't know if this suffices thought.

In the meanwhile I have one (kind'a silly) question: We referred to this package so far as Python KM, but have used more extensively `KramersMoyal`. I was wondering if it would make sense to resolve this somewhat ambiguous mentioning and simply refer to is as `KramersMoyal` and remove entirely the Python KM?
(This would imply also altering the title of the paper to "KramersMoyal: Kramers–Moyal coefficients for stochastic processes in python")

I find KM a bit too short and ambiguous (km as in kilometers...). E.g, MC for Monte Carlo is fairly used, but KM for Kramers–Moyal is not...

Best regards,
Leonardo Rydin
@fmeirinhos
