### 2019-12-01 6830017
Extending the docs, trying relative paths to git's root
