+++ docs/source/index.rst
@@ -1,19 +1,14 @@
-Guidelines to use KramersMoyal
-==============================
+Kramers---Moyal
+===============
 
 .. toctree::
    :maxdepth: 2
 
-   license
-
-KramersMoyal
-============
-
-Python KM is a python package designed to obtain the Kramers–Moyal
+Python KM is a python package designed to obtain the Kramers---Moyal
 coefficients, or conditional moments, from stochastic data of any
 dimension. It employs kernel density estimations, instead of a histogram
 approach, to ensure better results for low number of points as well as
-allowing better fitting of the results
+allowing better fitting of the results.
 
 Installation
 ============
@@ -28,7 +23,7 @@ Then on your favourite editor just use
 
 .. code:: python
 
-   from kramersmoyal import km, kernels
+   from kramersmoyal import km
 
 From here you can simply call
 
@@ -43,7 +38,7 @@ From here you can simply call
    powers = np.array([[1], [2]])
 
    # And here x is your (1D, 2D, 3D) data
-   edge, kmc = km(x, bins = bins, powers = powers)
+   kmc, edge = km(x, bins = bins, powers = powers)
 
 
 The library depends on ``numpy`` and ``scipy``.
@@ -55,24 +50,232 @@ A one-dimensional stochastic process
 The theory
 ----------
 
-Take for example the well documented one-dimension `Ornstein–Uhlenbeck <https://en.wikipedia.org/wiki/Ornstein%E2%80%93Uhlenbeck_process>`_
+Take, for example, the well-documented one-dimension `Ornstein---Uhlenbeck <https://en.wikipedia.org/wiki/Ornstein%E2%80%93Uhlenbeck_process>`_
 process, also known as Vašíček process.
 This process is governed by two main parameters: the mean-reverting
-parameter :math:`\theta` and the diffusion parameter :math:`\sigma`
-
+parameter :math:`\theta` and the diffusion or volatility coefficient :math:`\sigma`
 
 .. math::
    \mathrm{d}y(t) = -\theta y(t)\mathrm{d}t + \sigma \mathrm{d}W(t)
 
 which can be solved in various ways. For our purposes, recall that the drift
-coefficient, i.e., the first-order Kramers–Moyal coefficient, is given by
-:math:`\mathcal{M}^{[1]}(y) = \theta y` and the second-order Kramers–Moyal
+coefficient, i.e., the first-order Kramers---Moyal coefficient, is given by
+:math:`\mathcal{M}^{[1]}(y) = -\theta y` and the second-order Kramers---Moyal
 coefficient is :math:`\mathcal{M}^{[2]}(y) = \sigma^2 / 2`, i.e., the diffusion.
 
+For this example let's take :math:`\theta=0.3` and :math:`\sigma=0.1`, over a total time of 500
+units, with a sampling of 1000 Hertz, and from the generated data series
+retrieve the two parameters, the drift :math:`-\theta y(t)` and diffusion :math:`\sigma`.
 
-Indices and tables
-==================
 
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
+Integrating an Ornstein---Uhlenbeck process
+-------------------------------------------
+Here is a short code on generating a Ornstein---Uhlenbeck stochastic trajectory
+with a simple Euler–Maruyama integration method
+
+.. code:: python
+
+   # integration time and time sampling
+   t_final = 500
+   delta_t = 0.001
+
+   # The parameters theta and sigma
+   theta = 0.3
+   sigma = 0.1
+
[…]
