### 2019-12-01 e6cfe5b
Adding images for docs
