### 2019-12-02 LRydin/KramersMoyal#7 @kthyng
@Shibabrat Can you check on what @LRydin has added?
