### 2019-12-02 #1693 @kthyng
@LRydin I've pinged @Shibabrat on your issue to look at your work.

You can change the name of your package if you would like, including in the title of the paper, etc. Just go ahead and make all the changes if you want to do so.
