### 2019-12-02 #1693 @LRydin
Dear @kthyng,

wonderful, I'll make the changes then. I'm rereading the paper to ensure everything is okay. I'll commit the corrected version with the "new" name as soon as it is ready.

Best,
Leonardo Rydin
@fmeirinhos
