### 2019-12-03 LRydin/KramersMoyal#7 @Shibabrat
@kthyng looks good. I'm happy with this submission.

@fmeirinhos Yes, the readthedocs style documentation is certainly a good way to add more helpful examples and to introduce the theory for beginners.
