### 2019-12-03 98f1d85
Renaming the package, proof of the paper, 2 references added, corrected figures
