+++ paper/paper.md
@@ -1,7 +1,8 @@
 ---
-title: 'Python KM: Kramers--Moyal coefficients for stochastic processes'
+title: 'kramersmoyal: Kramers--Moyal coefficients for stochastic processes'
 tags:
   - Kramers--Moyal coefficents
+  - Kernel-density estimators
   - N-dimensional stochastic processes
   - Markovian processes
   - Master equation
@@ -17,15 +18,15 @@ authors:
 affiliations:
  - name: Department of Epileptology, University of Bonn, Venusberg Campus 1, 53127 Bonn, Germany
    index: 1
- - name: Helmholtz Institute for Radiation and Nuclear Physics, University of Bonn, Nußallee 14--16, 53115~Bonn, Germany
+ - name: Helmholtz Institute for Radiation and Nuclear Physics, University of Bonn, Nußallee 14--16, 53115 Bonn, Germany
    index: 2
- - name: Institute for Theoretical Physics, University of Cologne, 50937~Köln, Germany
+ - name: Forschungszentrum Jülich, Institute for Energy and Climate Research - Systems Analysis and Technology Evaluation (IEK-STE), 52428 Jülich, Germany
    index: 3
- - name: Forschungszentrum Jülich, Institute for Energy and Climate Research - Systems Analysis and Technology Evaluation (IEK-STE), 52428~Jülich, Germany
+ - name: Institute for Theoretical Physics, University of Cologne, 50937 Köln, Germany
    index: 4
  - name: Physikalisches Institut and Bethe Center for Theoretical Physics, Universität Bonn, Nussallee 12, 53115 Bonn, Germany
    index: 5
-date: 23rd of August, 2019
+date: 3rd of December, 2019
 bibliography: bib.bib
 ---
 
@@ -38,24 +39,24 @@ Given a set of stochastic data, ergodic or quasi-stationary, the extensive liter
 Most commonly known is the Fokker--Planck equation or truncated forward Kolmogorov equation, partial differential equations, obtained from the Taylor expansion of the master equation.
 
 Of particular relevance is the growing evidence that real-world data displays higher-order ($n>2$) Kramers--Moyal coefficients, which has a two-fold consequence: The common truncation at third order of the forward Kolmogorov equation, giving rise to the Fokker--Planck equation, is no longer valid.
-The evidence of higher-order ($n>2$) Kramers--Moyal coefficients in recorded data thus invalidates the aforementioned common argument for truncation, based on Pawula's theorem and thus rendering the Fokker--Planck description unjustified [@Tabar].
-A clear and common example is the presence of discontinuous jumps in data [@Anvari; @Sahalia], which can give rise to higher-order Kramers--Moyal coefficients, and are evidenced in @Rydin and references within.
+The existence of higher-order ($n>2$) Kramers--Moyal coefficients in recorded data thus invalidates the aforementioned common argument for truncation, thus rendering the Fokker--Planck description insufficient [@Tabar].
+A clear and common example is the presence of discontinuous jumps in data [@Anvari; @Sahalia], which can give rise to higher-order Kramers--Moyal coefficients, as are evidenced in @Rydin and references within.
 
 
 Calculating the moments or Kramers--Moyal coefficients stricly from data can be computationally heavy for long data series and is prone to be innacurate especially where the density of data points is scarce, e.g. usually at the boundaries on the domain of the process.
[…]
