+++ paper/bib.bib
@@ -65,7 +65,7 @@
   pages     = {043116},
   year      = {2018},
   doi       = {10.1063/1.5019320},
-  URL       = {https://doi.org/10.1063/1.5019320}
+  url       = {https://doi.org/10.1063/1.5019320}
 }
 
 @article{Sahalia,
@@ -76,7 +76,7 @@
   volume    = {57},
   pages     = {2075--2112},
   doi       = {10.1111/1540-6261.00489},
-  URL       = {https://doi.org/10.1111/1540-6261.00489}
+  url       = {https://doi.org/10.1111/1540-6261.00489}
 }
 
 @article{Rydin,
@@ -84,7 +84,7 @@
   author    = {Gorj{\~a}o, L. R. and Heysel, J. and Lehnertz, K. and Tabar, M. R. R.},
   journal   = {arXiv preprint arXiv:1907.05371},
   year      = {2019},
-  URL       = {https://arxiv.org/abs/1907.05371}
+  url       = {https://arxiv.org/abs/1907.05371}
 }
 
 @article{rinn2016,
@@ -103,7 +103,7 @@
   year      = {2018},
   publisher = {Routledge, New York},
   doi       = {10.1201/9781315140919},
-  ISBN      = {9781315140919}
+  isbn      = {9781315140919}
 }
 
 @book{Tabar,
@@ -123,4 +123,29 @@
   publisher = {Springer, Berlin},
   doi       = {10.1007/978-3-642-61544-3},
   isbn      = {978-3-540-61530-9}
+}
+
+@article{Nadaraya,
+  author    = {Nadaraya, E. A.},
+  title     = {On Estimating Regression},
+  journal   = {Theory of Probability \& Its Applications},
+  volume    = {9},
+  number    = {1},
+  pages     = {141-142},
+  year      = {1964},
+  doi       = {10.1137/1109020},
+  url       = {https://doi.org/10.1137/1109020},
+}
+
+@article{Watson,
+  author    = {Watson, Geoffrey S.},
+  title     = {Smooth Regression Analysis},
+  journal   = {Sankhy{\=a}: The Indian Journal of Statistics, Series A},
+  number    = {4},
+  pages     = {359--372},
+  publisher = {Springer},
+  volume    = {26},
+  year      = {1964},
+  url       = {http://www.jstor.org/stable/25049340},
+  issn      = {0581572X}
 }
