### 2019-12-03 420c721
Last corrections. Added funding for stay in Oslo.
