### 2019-12-03 6625735
changing the naming to kramersmoyal in docs and readme
