### 2019-12-04 #1693 @LRydin
Dear @kthyng, @Shibabrat, and @dawbarton,

I have made all the changes I deemed necessary. I reread the paper and made the final corrections. In this manner, I have only two remarks:
 - I've added the explicit mention of what the kernel-based estimations is: The [Nadaraya–Watson kernel regression](https://en.wikipedia.org/wiki/Kernel_regression), and the two relevant references there from Nadaraya and Watson.
- I've resized the figures and altered the name to `kramersmoyal`, as is both the name of the package and the GitHub repository.

There is one last issue: the paper *Analysis and data-driven reconstruction of bivariate jump-diffusion processes* that is referenced as being in the ArXiv is now in production. It was accepted last week. I believe it would be positive to add the actual reference (once the journal assigns an issue and number). @kthyng How should we proceed in this regard? Await for the issue number or publish with the reference to the ArXiv?

I think that should be it. Thank you all for the corrections, reviews and suggestions.

Best,
Leonardo Rydin
@fmeirinhos
