### 2019-12-04 #1693 @kthyng
@LRydin I think it makes sense to provide the most up to date information available when we're ready to finalize everything in your submission. Alternatively, if you prefer, we can wait for the full reference info on that paper and pause this in the meantime.
