### 2019-12-04 #1693 @LRydin
> @LRydin I think it makes sense to provide the most up to date information available when we're ready to finalize everything in your submission. Alternatively, if you prefer, we can wait for the full reference info on that paper and pause this in the meantime.

Dear @kthyng, I have reached out to the publisher to see if they already attributed an issue, but I have to wait for the publication procedure to advance a bit more. With luck next week there will be some more details, and then I could change the details in the JOSS paper. Else, I'd say, next week we finalise this, if all the boxes are ticked.

Best regards,
Leonardo Rydin
@fmeirinhos
