### 2019-12-06 LRydin/KramersMoyal#5 @kthyng
@dawbarton Is this issue ready to be closed from your perspective?
