### 2019-12-07 LRydin/KramersMoyal#5 @dawbarton
Yes, looks good to me.
