### 2019-12-12 #1693 @LRydin
Dear @kthyng,

I think we should proceed. I'll stick with the ArXiv version of the paper, as in the end the doi of the final publication is likely to also be linked to the same (ArXiv) doi. Should I alter the reference and include *in press* and the journal's name (Phys. Rev. E.)?

On another note, thank you @dawbarton and @Shibabrat for the reviews, the paper and the library are certainly looking better! I'm excited about my first JOSS publication.

Best,
Leonardo Rydin
@fmeirinhos
