### 2019-12-12 c893ea8
Adding JOSS badge for the paper :)
