### 2019-12-12 cce5a3a
Update to bib
