+++ paper/bib.bib
@@ -70,8 +70,8 @@
 
 @article{Sahalia,
   title     = {Telling from discrete data whether the underlying continuous-time model is a diffusion},
-  author    = {A{\"\i}t-Sahalia, Y.},
-  journal   = {J. Finance},
+  author    = {A{\"i}t-Sahalia, Y.},
+  journal   = {Journal Finance},
   year      = {2002},
   volume    = {57},
   pages     = {2075--2112},
@@ -82,7 +82,8 @@
 @article{Rydin,
   title     = {Analysis and data-driven reconstruction of bivariate jump-diffusion processes},
   author    = {Gorj{\~a}o, L. R. and Heysel, J. and Lehnertz, K. and Tabar, M. R. R.},
-  journal   = {arXiv preprint arXiv:1907.05371},
+  journal   = {Physical Review E},
+  volume    = {(in press)},
   year      = {2019},
   url       = {https://arxiv.org/abs/1907.05371}
 }
