### 2019-12-12 a2fb3dc
Corrected name of journal in bib
