+++ paper/bib.bib
@@ -71,7 +71,7 @@
 @article{Sahalia,
   title     = {Telling from discrete data whether the underlying continuous-time model is a diffusion},
   author    = {A{\"i}t-Sahalia, Y.},
-  journal   = {Journal Finance},
+  journal   = {The Journal of Finance},
   year      = {2002},
   volume    = {57},
   pages     = {2075--2112},
