### 2019-12-12 #1693 @LRydin
Dear @kthyng,

That's that. It is ready. The last thing to do is to get the figures out of the tables and simply utilise captions... which I have no idea how it's done... Do you do that on your side, in the processing for the publication?

Best regards,
Leonardo Rydin
@fmeirinhos
