### 2019-12-16 4f7486b
Changing setup for figures
Going to test this in the preview tool next.
