### 2019-12-16 #1693 @kthyng
@LRydin I have edited the figures in your paper in PR [#8](https://github.com/LRydin/KramersMoyal/pull/8). If you think the paper looks good with the changes, you can accept the PR.
