### 2019-12-17 2d5b6d9
Removed comments on paper. Added space before figure
