### 2019-12-17 #1693 @LRydin
Dear @kthyng,

I've accepted the PR, but also added a left and right margin to the figures, to have them all in the same size. I've updated the repo. I think that should be that :).

Cheers,
Leonardo Rydin
@fmeirinhos
