### 2019-12-17 418f80c
Adding badges
