### 2019-12-17 1685433
Lots of small changes.
I made lots of typo fixes and small wording changes. Please check them in detail. Additionally, there is a word "riddance" still in the paper that isn't a word I know but could be a technical word here. If it isn't, please replace.
