+++ paper/paper.md
@@ -44,22 +44,22 @@ The existence of higher-order ($n>2$) Kramers--Moyal coefficients in recorded da
 A clear and common example is the presence of discontinuous jumps in data [@Anvari; @Sahalia], which can give rise to higher-order Kramers--Moyal coefficients, as are evidenced in @Rydin and references within.
 
 
-Calculating the moments or Kramers--Moyal coefficients stricly from data can be computationally heavy for long data series and is prone to be innacurate especially where the density of data points is scarce, e.g. usually at the boundaries on the domain of the process.
+Calculating the moments or Kramers--Moyal coefficients strictly from data can be computationally heavy for long data series and is prone to innaccuracy especially where the density of data points is scarce, for example, usually at the boundaries on the domain of the process.
 The most straightforward approach is to perform a histogram-based estimation to evaluate the moments of the system at hand.
 This has two main drawbacks: it requires a discrete space of examination of the process and is shown to be less accurate than using kernel-based estimators [@Lamouroux].
 
-This library is based on a kernel-based estimation, i.e., the Nadaraya--Watson kernel estimator [@Nadaraya; @Watson], which allows for more robust results given both a wider range of possible kernel shapes to perform the calculation, as well as retrieving the results in a non-binned coordinate space, unlike histogram regressions [@Silverman].
+This library is based on a kernel-based estimation, *i.e.*, the Nadaraya--Watson kernel estimator [@Nadaraya; @Watson], which allows for more robust results given both a wider range of possible kernel shapes to perform the calculation, as well as retrieving the results in a non-binned coordinate space, unlike histogram regressions [@Silverman].
 It further employs a convolution of the time series with the selected kernel, circumventing the computational issue of sequential array summation, the most common bottleneck in integration time and computer memory.
 
-The package presented here contains several options: A general open-source toolbox for the calculation of Kramers--Moyal coefficients for any given data series of any dimension and to any order, with a selection of commonly used kernel estimators.
+The package presented here contains several options: A general open-source toolbox for the calculation of Kramers--Moyal coefficients for any given data series of any dimension and to any order, with a selection of commonly-used kernel estimators.
 
 # Mathematics
 
-For a general $N$-dimensional Markovian process $\boldsymbol{x}(t)\in\mathbb{R}^N$ the Kramers--Moyal yields all orders of the cumulants of the conditional probability distribution $P(\boldsymbol{x}',t+\Delta T | \boldsymbol{x}, t)$ as [@Risken]
+For a general $N$-dimensional Markovian process $\boldsymbol{x}(t)\in\mathbb{R}^N$ the Kramers--Moyal yields all orders of the cumulants of the conditional probability distribution $P(\boldsymbol{x}',t+\Delta T | \boldsymbol{x}, t)$ as
 $$
 \mathcal{M}^{\sigma}(\boldsymbol{x},t)=\lim_{\Delta t\to 0}\frac{1}{\Delta t}\int  d\boldsymbol{x}'[\boldsymbol{x}(t)'-\boldsymbol{x}(t)]^\sigma P(\boldsymbol{x}',t+\Delta T | \boldsymbol{x}, t),\nonumber \tag*{(1)}
 $$
-with $[\dots]^\sigma$ a dyadic multiplication and the power $\sigma$ allowing for a set of powers depending on the dimensionality of the process.
+with $[\dots]^\sigma$ a dyadic multiplication and the power $\sigma$ allowing for a set of powers depending on the dimensionality of the process [@Risken].
 
[…]
