### 2019-12-17 #1693 @kthyng
I have made some correction for your paper in this PR: https://github.com/LRydin/KramersMoyal/pull/9
