### 2019-12-17 #1693 @LRydin
Dear @kthyng,

> We recommend making an official release for your submission. I see we have v0.31 marked. I want to find out 1. if that is still the correct version number and 2. if it is, are you interested in making a corresponding git release to match?

1. The current version is at 0.4
2. Zenodo: Done. Here is the link: https://zenodo.org/record/3581328#.XfkWPdnTWbd and the DOI is 10.21105/joss.01693

I have accepted the last PR and corrected the suggested spellings.

Cheers,
Leonardo
@fmeirinhos
