### 2019-12-18 #1693 @kthyng
Are you interested in making a formal github release, such that it would show up here: https://github.com/LRydin/KramersMoyal/releases

You can get more info here: https://help.github.com/en/github/administering-a-repository/creating-releases
