### 2019-12-18 #1693 @LRydin
Seems like a good idea @kthyng. Should I do so now?
