### 2019-12-18 #1693 @LRydin
@kthyng Here it is: https://github.com/LRydin/KramersMoyal/releases/tag/v0.4
