### 2019-12-19 #1693 @kthyng
Congrats on your new paper @LRydin!!! Thanks to  reviewers @dawbarton and @Shibabrat — we couldn't have done this without your time and expertise.
