### 2019-12-19 #1693 @LRydin
Exciting! Thanks @kthyng, @dawbarton and @Shibabrat. I will proceed to do
> Party like you just published a paper!

as expected :)

Thank you all again.
Cheers,
Leonardo Rydin
@fmeirinhos
